coolpress/requirements.txt
djangorestframework
libgravatar
requests
wordcloud
BeautifulSoup4